DROID-3819 fixes
