+++ app/src/main/java/com/anytypeio/anytype/ui/vault/VaultScreen.kt
@@ -1,516 +1,55 @@
 package com.anytypeio.anytype.ui.vault
 
-import android.content.res.Configuration
 import android.os.Build.VERSION.SDK_INT
 import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.systemBars
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.Scaffold
-import androidx.compose.material.Text
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.rememberNestedScrollInteropConnection
 import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
-import androidx.constraintlayout.compose.ConstraintLayout
-import coil3.compose.rememberAsyncImagePainter
 import com.anytypeio.anytype.R
 import com.anytypeio.anytype.core_models.Id
 import com.anytypeio.anytype.core_ui.common.ReorderHapticFeedbackType
 import com.anytypeio.anytype.core_ui.common.rememberReorderHapticFeedback
-import com.anytypeio.anytype.core_ui.foundation.DefaultSearchBar
-import com.anytypeio.anytype.core_ui.foundation.Divider
-import com.anytypeio.anytype.core_ui.foundation.noRippleThrottledClickable
-import com.anytypeio.anytype.core_ui.views.BodyRegular
-import com.anytypeio.anytype.core_ui.views.PreviewTitle2Regular
-import com.anytypeio.anytype.core_ui.views.Title1
-import com.anytypeio.anytype.core_ui.widgets.ListWidgetObjectIcon
 import com.anytypeio.anytype.core_utils.insets.EDGE_TO_EDGE_MIN_SDK
-import com.anytypeio.anytype.presentation.objects.ObjectIcon.Profile.Avatar
 import com.anytypeio.anytype.presentation.profile.AccountProfile
-import com.anytypeio.anytype.presentation.profile.ProfileIconView
-import com.anytypeio.anytype.presentation.spaces.SelectSpaceViewModel
-import com.anytypeio.anytype.presentation.vault.VaultSectionView
 import com.anytypeio.anytype.presentation.vault.VaultSpaceView
+import com.anytypeio.anytype.presentation.vault.VaultUiState
 import kotlinx.coroutines.delay
 import sh.calvin.reorderable.ReorderableItem
 import sh.calvin.reorderable.rememberReorderableLazyListState
 
-
 @Composable
-fun VaultScreenToolbar(
+fun VaultScreen(
     profile: AccountProfile,
-    spaceCountLimitReached: Boolean = false,
-    showNotificationBadge: Boolean = false,
-    onPlusClicked: () -> Unit,
-    onSettingsClicked: () -> Unit,
-    isLoading: Boolean
-) {
-
-    Column {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(52.dp)
-        ) {
-            Box(
-                modifier = Modifier
-                    .width(64.dp)
-                    .fillMaxHeight()
-                    .noRippleThrottledClickable {
-                        onSettingsClicked()
-                    },
-                contentAlignment = Alignment.Center
-
-            ) {
-                ProfileIconWithBadge(
-                    modifier = Modifier.size(34.dp),
-                    profile = profile,
-                    showBadge = showNotificationBadge
-                )
-            }
-
-            ConstraintLayout(
-                modifier = Modifier.fillMaxSize(),
-            ) {
-                val (t, l) = createRefs()
-                Text(
-                    modifier = Modifier
-                        .wrapContentWidth()
-                        .constrainAs(t) {
-                            start.linkTo(parent.start)
-                            end.linkTo(parent.end)
-                            top.linkTo(parent.top)
-                            bottom.linkTo(parent.bottom)
-                        }
-                    ,
-                    style = Title1,
-                    text = stringResource(R.string.vault_my_spaces),
-                    color = colorResource(id = R.color.text_primary),
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-                if (isLoading) {
-                    CircularProgressIndicator(
-                        modifier = Modifier
-                            .padding(end = 7.dp)
-                            .size(16.dp)
-                            .constrainAs(l) {
-                                end.linkTo(t.start)
-                                top.linkTo(parent.top)
-                                bottom.linkTo(parent.bottom)
-                            },
-                        color = colorResource(id = R.color.glyph_active),
-                        strokeWidth = 2.dp
-                    )
-                }
-            }
-
-            if (!spaceCountLimitReached) {
-                Box(
-                    modifier = Modifier
-                        .align(Alignment.CenterEnd)
-                        .width(64.dp)
-                        .fillMaxHeight()
-                        .noRippleThrottledClickable {
-                            onPlusClicked()
-                        },
-                    contentAlignment = Alignment.Center
-
-                ) {
-                    Image(
-                        painter = painterResource(id = R.drawable.ic_plus_18),
-                        contentDescription = stringResource(R.string.content_description_plus_button),
-                        modifier = Modifier.size(28.dp),
-                        contentScale = ContentScale.Fit
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ProfileIcon(
-    modifier: Modifier,
-    profile: AccountProfile
-) {
-    when (profile) {
-        is AccountProfile.Data -> {
-            Box(modifier) {
-                when (val icon = profile.icon) {
-                    is ProfileIconView.Image -> {
-                        Image(
-                            painter = rememberAsyncImagePainter(icon.url),
-                            contentDescription = "Custom image profile",
-                            contentScale = ContentScale.Crop,
-                            modifier = Modifier
-                                .fillMaxSize()
-                                .clip(CircleShape)
-                        )
-                    }
-
-                    else -> {
-                        val nameFirstChar = if (profile.name.isEmpty()) {
-                            stringResource(id = R.string.account_default_name)
-                        } else {
-                            profile.name.first().uppercaseChar().toString()
-                        }
-                        ListWidgetObjectIcon(
-                            modifier = Modifier.fillMaxSize(),
-                            icon = Avatar(
-                                name = nameFirstChar
-                            ),
-                            iconSize = 28.dp
-                        )
-                    }
-                }
-            }
-        }
-
-        AccountProfile.Idle -> {
-            // Draw nothing
-        }
-    }
-}
-
-@Composable
-private fun ProfileIconWithBadge(
-    modifier: Modifier,
-    profile: AccountProfile,
-    showBadge: Boolean = false
-) {
-    Box(
-        modifier = modifier
-    ) {
-        // Main profile icon
-        ProfileIcon(
-            modifier = Modifier
-                .size(28.dp)
-                .align(Alignment.Center),
-            profile = profile
-        )
-
-        // Badge positioned in top-right corner
-        if (showBadge) {
-            Box(
-                modifier = Modifier
-                    .align(Alignment.TopEnd)
-                    .size(12.dp)
-                    .background(
-                        color = colorResource(id = R.color.background_primary),
-                        shape = CircleShape
-                    ),
-                contentAlignment = Alignment.Center,
-            ) {
-                Spacer(
-                    modifier = Modifier
-                        .size(8.dp)
-                        .background(
-                            color = colorResource(id = R.color.palette_system_red),
-                            shape = CircleShape
-                        )
-                )
-            }
-        }
-    }
-}
-
-@Composable
-@Preview(
-    showBackground = true,
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    name = "Dark Mode - Not Scrolled"
-)
-@Preview(
-    showBackground = true,
-    uiMode = Configuration.UI_MODE_NIGHT_NO,
-    name = "Light Mode - Not Scrolled"
-)
-fun VaultScreenToolbarNotScrolledPreview() {
-    VaultScreenToolbar(
-        onPlusClicked = {},
-        onSettingsClicked = {},
-        profile = AccountProfile.Data(
-            name = "John Doe",
-            icon = ProfileIconView.Placeholder(name = "Jd")
-        ),
-        isLoading = false,
-    )
-}
-
-@Composable
-@Preview(
-    showBackground = true,
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    name = "Dark Mode - Scrolled"
-)
-@Preview(
-    showBackground = true,
-    uiMode = Configuration.UI_MODE_NIGHT_NO,
-    name = "Light Mode - Scrolled"
-)
-fun VaultScreenToolbarScrolledPreview() {
-    VaultScreenToolbar(
-        onPlusClicked = {},
-        onSettingsClicked = {},
-        profile = AccountProfile.Data(
-            name = "John Doe",
-            icon = ProfileIconView.Placeholder(name = "Jd")
-        ),
-        isLoading = false,
-        showNotificationBadge = true
-    )
-}
-
-@Composable
-fun SpaceActionsDropdownMenu(
-    expanded: Boolean,
-    onDismiss: () -> Unit,
-    isMuted: Boolean?,
-    isPinned: Boolean,
-    maxPinnedSpaces: Int,
-    showPinButton: Boolean,
-    onMuteToggle: () -> Unit,
-    onPinToggle: () -> Unit,
-    onSpaceSettings: () -> Unit
-) {
-    DropdownMenu(
-        modifier = Modifier.width(254.dp),
-        expanded = expanded,
-        onDismissRequest = onDismiss,
-        containerColor = colorResource(R.color.background_secondary),
-        shape = RoundedCornerShape(12.dp),
-        tonalElevation = 8.dp,
-        offset = DpOffset(
-            x = 16.dp,
-            y = 8.dp
-        )
-    ) {
-        if (!showPinButton && !isPinned) {
-            // Show info message instead of Pin
-            Row(modifier = Modifier.padding(16.dp)) {
-                Text(
-                    text = stringResource(R.string.vault_pinned_limit_message, maxPinnedSpaces),
-                    style = PreviewTitle2Regular,
-                    color = colorResource(id = R.color.text_secondary)
-                )
-            }
-        } else {
-            DropdownMenuItem(
-                onClick = {
-                    onPinToggle()
-                    onDismiss()
-                },
-                text = {
-                    val stringRes = if (isPinned) {
-                        R.string.vault_unpin_space
-                    } else {
-                        R.string.vault_pin_space
-                    }
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text(
-                            modifier = Modifier.weight(1f),
-                            text = stringResource(id = stringRes),
-                            style = BodyRegular,
-                            color = colorResource(id = R.color.text_primary)
-                        )
-                        Image(
-                            painter = painterResource(id = if (isPinned) R.drawable.ic_unpin_24 else R.drawable.ic_pin_24),
-                            contentDescription = null,
-                            modifier = Modifier
-                                .padding(end = 4.dp)
-                                .size(24.dp)
-                        )
-                    }
-                }
-            )
-        }
-
-        // Mute/Unmute only if chat enabled
-        if (isMuted != null) {
-            Divider(paddingStart = 0.dp, paddingEnd = 0.dp)
-            DropdownMenuItem(
-                onClick = {
-                    onMuteToggle()
-                    onDismiss()
-                },
-                text = {
-                    val (stringRes, iconRes) = if (isMuted) {
-                        R.string.space_notify_unmute to R.drawable.ic_notifications
-                    } else {
-                        R.string.space_notify_mute to R.drawable.ic_notifications_off
-                    }
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text(
-                            modifier = Modifier.weight(1f),
-                            text = stringResource(id = stringRes),
-                            style = BodyRegular,
-                            color = colorResource(id = R.color.text_primary)
-                        )
-                        Image(
-                            painter = painterResource(id = iconRes),
-                            contentDescription = null,
-                            modifier = Modifier
-                                .padding(end = 4.dp)
-                                .size(24.dp)
-                        )
-                    }
-                }
-            )
-        }
-        Divider(paddingStart = 0.dp, paddingEnd = 0.dp)
-        // Space Settings (always shown)
-        DropdownMenuItem(
-            onClick = {
-                onSpaceSettings()
-                onDismiss()
-            },
-            text = {
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Text(
-                        modifier = Modifier.weight(1f),
-                        style = BodyRegular,
-                        color = colorResource(id = R.color.text_primary),
-                        text = stringResource(R.string.vault_space_settings)
-                    )
-                    Image(
-                        painter = painterResource(id = R.drawable.ic_space_settings_24),
-                        contentDescription = null,
-                        modifier = Modifier
-                            .padding(end = 4.dp)
-                            .size(24.dp)
-                    )
-                }
-            }
-        )
-    }
-}
-
-@Preview(showBackground = true, name = "SpaceActionsDropdownMenu - Muted Owner")
-@Composable
-fun PreviewSpaceActionsDropdownMenu_MutedOwner() {
-    var expanded by remember { mutableStateOf(true) }
-    Box(Modifier.fillMaxSize()) {
-        SpaceActionsDropdownMenu(
-            expanded = expanded,
-            onDismiss = { expanded = false },
-            isMuted = true,
-            isPinned = false,
-            maxPinnedSpaces = VaultSectionView.MAX_PINNED_SPACES,
-            showPinButton = true,
-            onMuteToggle = {},
-            onPinToggle = {},
-            onSpaceSettings = {}
-        )
-    }
-}
-
-@Preview(showBackground = true, name = "SpaceActionsDropdownMenu - Unmuted Not Owner")
-@Composable
-fun PreviewSpaceActionsDropdownMenu_UnmutedNotOwner() {
-    var expanded by remember { mutableStateOf(true) }
-    Box(Modifier.fillMaxSize()) {
-        SpaceActionsDropdownMenu(
-            expanded = expanded,
-            onDismiss = { expanded = false },
-            isMuted = false,
-            isPinned = false,
-            maxPinnedSpaces = VaultSectionView.MAX_PINNED_SPACES,
-            showPinButton = true,
-            onMuteToggle = {},
-            onPinToggle = {},
-            onSpaceSettings = {}
-        )
-    }
-}
-
-@Composable
-fun SpaceActionsDropdownMenuHost(
-    spaceView: VaultSpaceView,
-    expanded: Boolean,
-    onDismiss: () -> Unit,
-    onMuteSpace: (Id) -> Unit,
-    onUnmuteSpace: (Id) -> Unit,
-    onPinSpace: (Id) -> Unit,
-    onUnpinSpace: (Id) -> Unit,
-    maxPinnedSpaces: Int,
-    showPinButton: Boolean,
-    onSpaceSettings: (Id) -> Unit
-) {
-    SpaceActionsDropdownMenu(
-        expanded = expanded,
-        onDismiss = onDismiss,
-        isMuted = spaceView.isMuted,
-        isPinned = spaceView.isPinned,
-        showPinButton = showPinButton,
-        maxPinnedSpaces = maxPinnedSpaces,
-        onMuteToggle = {
-            spaceView.space.targetSpaceId?.let {
-                if (spaceView.isMuted == true) onUnmuteSpace(it) else onMuteSpace(it)
-            }
-        },
-        onPinToggle = {
-            spaceView.space.id.let {
-                if (spaceView.isPinned) onUnpinSpace(it) else onPinSpace(it)
-            }
-        },
-        onSpaceSettings = {
-            spaceView.space.id.let { onSpaceSettings(it) }
-        }
-    )
-}
-
-@Composable
-fun VaultScreenWithUnreadSection(
-    profile: AccountProfile,
-    sections: VaultSectionView,
+    uiState: VaultUiState,
     showNotificationBadge: Boolean = false,
     onSpaceClicked: (VaultSpaceView) -> Unit,
     onCreateSpaceClicked: () -> Unit,
     onSettingsClicked: () -> Unit,
-    isLoading: Boolean,
     onMuteSpace: (Id) -> Unit,
     onUnmuteSpace: (Id) -> Unit,
     onPinSpace: (Id) -> Unit,
@@ -521,31 +60,75 @@ fun VaultScreenWithUnreadSection(
 ) {
 
     var searchQuery by remember { mutableStateOf("") }
-    val filteredMainSpaces = remember(searchQuery, sections.mainSpaces) {
-        if (searchQuery.isBlank()) {
-            sections.mainSpaces
-        } else {
-            sections.mainSpaces.filter { space ->
-                space.space.name?.contains(searchQuery, ignoreCase = true) == true
-            }
-        }
-    }
-
-    val filteredPinnedSpaces = remember(searchQuery, sections.pinnedSpaces) {
-        if (searchQuery.isBlank()) {
-            sections.pinnedSpaces
-        } else {
-            sections.pinnedSpaces.filter { space ->
-                space.space.name?.contains(searchQuery, ignoreCase = true) == true
-            }
-        }
-    }
-
-    val hasAnyFilteredSpaces = filteredMainSpaces.isNotEmpty() || filteredPinnedSpaces.isNotEmpty()
-
     val lazyListState = rememberLazyListState()
-    var expandedSpaceId by remember { mutableStateOf<String?>(null) }
 
+    Scaffold(
+        modifier = Modifier
+            .fillMaxSize()
+            .nestedScroll(rememberNestedScrollInteropConnection())
+            .background(color = colorResource(id = R.color.background_primary))
+            .then(
+                if (SDK_INT >= EDGE_TO_EDGE_MIN_SDK)
+                    Modifier.windowInsetsPadding(WindowInsets.systemBars)
+                else
+                    Modifier
+            ),
+        backgroundColor = colorResource(id = R.color.background_primary),
+        topBar = {
+            VaultScreenTopToolbar(
+                profile = profile,
+                uiState = uiState,
+                searchQuery = searchQuery,
+                showNotificationBadge = showNotificationBadge,
+                onCreateSpaceClicked = onCreateSpaceClicked,
+                onSettingsClicked = onSettingsClicked,
+                onUpdateSearchQuery = { query ->
+                    searchQuery = query
+                }
+            )
+        }
+    ) { paddings ->
+        when (uiState) {
+            VaultUiState.Loading -> {}
+
+            is VaultUiState.Sections -> {
+                VaultScreenContent(
+                    sections = uiState,
+                    lazyListState = lazyListState,
+                    paddings = paddings,
+                    searchQuery = searchQuery,
+                    onSpaceClicked = onSpaceClicked,
+                    onCreateSpaceClicked = onCreateSpaceClicked,
+                    onMuteSpace = onMuteSpace,
+                    onUnmuteSpace = onUnmuteSpace,
+                    onPinSpace = onPinSpace,
+                    onUnpinSpace = onUnpinSpace,
+                    onOrderChanged = onOrderChanged,
+                    onDragEnd = onDragEnd,
+                    onSpaceSettings = onSpaceSettings
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun VaultScreenContent(
+    sections: VaultUiState.Sections,
+    lazyListState: LazyListState,
+    paddings: PaddingValues,
+    searchQuery: String,
+    onSpaceClicked: (VaultSpaceView) -> Unit,
+    onCreateSpaceClicked: () -> Unit,
+    onMuteSpace: (Id) -> Unit,
+    onUnmuteSpace: (Id) -> Unit,
+    onPinSpace: (Id) -> Unit,
+    onUnpinSpace: (Id) -> Unit,
+    onOrderChanged: (String, String) -> Unit,
+    onDragEnd: () -> Unit = { /* No-op */ },
+    onSpaceSettings: (Id) -> Unit
+) {
+    var expandedSpaceId by remember { mutableStateOf<String?>(null) }
 
     val hapticFeedback = rememberReorderHapticFeedback()
 
@@ -572,218 +155,96 @@ fun VaultScreenWithUnreadSection(
         }
     }
 
-    Scaffold(
-        modifier = Modifier
-            .fillMaxSize()
-            .nestedScroll(rememberNestedScrollInteropConnection())
-            .background(color = colorResource(id = R.color.background_primary))
-            .then(
-                if (SDK_INT >= EDGE_TO_EDGE_MIN_SDK)
-                    Modifier.windowInsetsPadding(WindowInsets.systemBars)
-                else
-                    Modifier
-            ),
-        backgroundColor = colorResource(id = R.color.background_primary),
-        topBar = {
-            Column {
-                VaultScreenToolbar(
-                    profile = profile,
-                    showNotificationBadge = showNotificationBadge,
-                    onPlusClicked = onCreateSpaceClicked,
-                    onSettingsClicked = onSettingsClicked,
-                    spaceCountLimitReached = sections.mainSpaces.size >= SelectSpaceViewModel.MAX_SPACE_COUNT,
-                    isLoading = isLoading
-                )
-                if (isLoading) {
-                    Spacer(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(60.dp)
-                    )
-                } else {
-                    DefaultSearchBar(
-                        value = searchQuery,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 16.dp, vertical = 10.dp),
-                        onQueryChanged = { query -> searchQuery = query }
-                    )
-                }
+    val filteredMainSpaces = remember(searchQuery, sections.mainSpaces) {
+        if (searchQuery.isBlank()) {
+            sections.mainSpaces
+        } else {
+            sections.mainSpaces.filter { space ->
+                space.space.name?.contains(searchQuery, ignoreCase = true) == true
             }
         }
-    ) { paddings ->
-        if (isLoading) {
-            Spacer(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(paddings),
-            )
-            return@Scaffold
-        }
-        if (sections.mainSpaces.isEmpty() && sections.pinnedSpaces.isEmpty()) {
-            VaultEmptyState(
-                modifier = Modifier.padding(paddings),
-                onCreateSpaceClicked = onCreateSpaceClicked
-            )
-        } else if (!hasAnyFilteredSpaces) {
-            VaultEmptyState(
-                modifier = Modifier.padding(paddings),
-                textRes = R.string.vault_empty_search_state_text,
-                showButton = false,
-                onCreateSpaceClicked = onCreateSpaceClicked
-            )
+    }
+
+    val filteredPinnedSpaces = remember(searchQuery, sections.pinnedSpaces) {
+        if (searchQuery.isBlank()) {
+            sections.pinnedSpaces
         } else {
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(paddings),
-                state = lazyListState,
-                verticalArrangement = Arrangement.spacedBy(8.dp),
-                contentPadding = PaddingValues(top = 4.dp)
-            ) {
-                // Pinned Spaces Section
-                if (filteredPinnedSpaces.isNotEmpty()) {
-                    itemsIndexed(
-                        items = filteredPinnedSpaces,
-                        key = { _, item -> item.space.id },
-                        contentType = { _, item ->
-                            when (item) {
-                                is VaultSpaceView.Chat -> TYPE_CHAT
-                                is VaultSpaceView.Space -> TYPE_SPACE
-                            }
-                        }
-                    ) { index, item ->
-                        ReorderableItem(
-                            state = reorderableLazyListState,
-                            key = item.space.id
-                        ) { isDragging ->
-                            val alpha = animateFloatAsState(if (isDragging) 0.8f else 1.0f)
+            sections.pinnedSpaces.filter { space ->
+                space.space.name?.contains(searchQuery, ignoreCase = true) == true
+            }
+        }
+    }
 
-                            when (item) {
-                                is VaultSpaceView.Chat -> {
-                                    VaultChatCard(
-                                        modifier = Modifier
-                                            .fillMaxWidth()
-                                            .height(80.dp)
-                                            .combinedClickable(
-                                                onClick = {
-                                                    onSpaceClicked(item)
-                                                },
-                                                onLongClick = {
-                                                    expandedSpaceId = item.space.id
-                                                }
-                                            )
-                                            .padding(horizontal = 16.dp)
-                                            .graphicsLayer(alpha = alpha.value)
-                                            .animateItem()
-                                            .longPressDraggableHandle(
-                                                onDragStarted = {
-                                                    hapticFeedback.performHapticFeedback(
-                                                        ReorderHapticFeedbackType.START
-                                                    )
-                                                },
-                                                onDragStopped = {
-                                                    hapticFeedback.performHapticFeedback(
-                                                        ReorderHapticFeedbackType.MOVE
-                                                    )
-                                                }
-                                            ),
-                                        title = item.space.name.orEmpty(),
-                                        icon = item.icon,
-                                        previewText = item.previewText,
-                                        creatorName = item.creatorName,
-                                        messageText = item.messageText,
-                                        messageTime = item.messageTime,
-                                        chatPreview = item.chatPreview,
-                                        unreadMessageCount = item.unreadMessageCount,
-                                        unreadMentionCount = item.unreadMentionCount,
-                                        attachmentPreviews = item.attachmentPreviews,
-                                        isMuted = item.isMuted,
-                                        isPinned = item.isPinned,
-                                        spaceView = item,
-                                        expandedSpaceId = expandedSpaceId,
-                                        onDismissMenu = { expandedSpaceId = null },
-                                        onMuteSpace = onMuteSpace,
-                                        onUnmuteSpace = onUnmuteSpace,
-                                        onPinSpace = onPinSpace,
-                                        onUnpinSpace = onUnpinSpace,
-                                        onSpaceSettings = onSpaceSettings,
-                                        maxPinnedSpaces = VaultSectionView.MAX_PINNED_SPACES,
-                                        showPinButton = item.showPinButton
-                                    )
-                                }
+    val hasAnyFilteredSpaces = filteredMainSpaces.isNotEmpty() || filteredPinnedSpaces.isNotEmpty()
 
-                                is VaultSpaceView.Space -> {
-                                    VaultSpaceCard(
-                                        modifier = Modifier
-                                            .animateItem()
-                                            .graphicsLayer(alpha = alpha.value)
-                                            .combinedClickable(
-                                                onClick = {
-                                                    onSpaceClicked(item)
-                                                },
-                                                onLongClick = {
-                                                    expandedSpaceId = item.space.id
-                                                }
-                                            )
-                                            .longPressDraggableHandle(
-                                                onDragStarted = {
-                                                    hapticFeedback.performHapticFeedback(
-                                                        ReorderHapticFeedbackType.START
-                                                    )
-                                                },
-                                                onDragStopped = {
-                                                    hapticFeedback.performHapticFeedback(
-                                                        ReorderHapticFeedbackType.MOVE
-                                                    )
-                                                }
-                                            ),
-                                        title = item.space.name.orEmpty(),
-                                        subtitle = item.accessType,
-                                        icon = item.icon,
-                                        isPinned = item.isPinned,
-                                        spaceView = item,
-                                        expandedSpaceId = expandedSpaceId,
-                                        onDismissMenu = { expandedSpaceId = null },
-                                        onMuteSpace = onMuteSpace,
-                                        onUnmuteSpace = onUnmuteSpace,
-                                        onPinSpace = onPinSpace,
-                                        onUnpinSpace = onUnpinSpace,
-                                        maxPinnedSpaces = VaultSectionView.MAX_PINNED_SPACES,
-                                        onSpaceSettings = onSpaceSettings,
-                                        showPinButton = item.showPinButton
-                                    )
-                                }
-                            }
+    if (sections.mainSpaces.isEmpty() && sections.pinnedSpaces.isEmpty()) {
+        VaultEmptyState(
+            modifier = Modifier.padding(paddings),
+            onCreateSpaceClicked = onCreateSpaceClicked
+        )
+    } else if (!hasAnyFilteredSpaces) {
+        VaultEmptyState(
+            modifier = Modifier.padding(paddings),
+            textRes = R.string.vault_empty_search_state_text,
+            showButton = false,
+            onCreateSpaceClicked = onCreateSpaceClicked
+        )
+    } else {
+        LazyColumn(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(paddings),
+            state = lazyListState,
+            verticalArrangement = Arrangement.spacedBy(8.dp),
+            contentPadding = PaddingValues(top = 4.dp)
+        ) {
+            // Pinned Spaces Section
+            if (filteredPinnedSpaces.isNotEmpty()) {
+                itemsIndexed(
+                    items = filteredPinnedSpaces,
+                    key = { _, item -> item.space.id },
+                    contentType = { _, item ->
+                        when (item) {
+                            is VaultSpaceView.Chat -> TYPE_CHAT
+                            is VaultSpaceView.Space -> TYPE_SPACE
                         }
                     }
-                }
+                ) { index, item ->
+                    ReorderableItem(
+                        state = reorderableLazyListState,
+                        key = item.space.id
+                    ) { isDragging ->
+                        val alpha = animateFloatAsState(if (isDragging) 0.8f else 1.0f)
 
-                // Not pinned Spaces Section
-                if (filteredMainSpaces.isNotEmpty()) {
-                    itemsIndexed(
-                        items = filteredMainSpaces,
-                        key = { _, item -> item.space.id },
-                        contentType = { _, item ->
-                            when (item) {
-                                is VaultSpaceView.Chat -> TYPE_CHAT
-                                is VaultSpaceView.Space -> TYPE_SPACE
-                            }
-                        }
-                    ) { _, item ->
                         when (item) {
                             is VaultSpaceView.Chat -> {
                                 VaultChatCard(
                                     modifier = Modifier
-                                        .animateItem()
                                         .fillMaxWidth()
                                         .height(80.dp)
-                                        .then(
-                                            createCombinedClickableModifier(
-                                                onClick = { onSpaceClicked(item) },
-                                                onLongClick = { expandedSpaceId = item.space.id }
-                                            ))
-                                        .padding(horizontal = 16.dp),
+                                        .combinedClickable(
+                                            onClick = {
+                                                onSpaceClicked(item)
+                                            },
+                                            onLongClick = {
+                                                expandedSpaceId = item.space.id
+                                            }
+                                        )
+                                        .padding(horizontal = 16.dp)
+                                        .graphicsLayer(alpha = alpha.value)
+                                        .animateItem()
+                                        .longPressDraggableHandle(
+                                            onDragStarted = {
+                                                hapticFeedback.performHapticFeedback(
+                                                    ReorderHapticFeedbackType.START
+                                                )
+                                            },
+                                            onDragStopped = {
+                                                hapticFeedback.performHapticFeedback(
+                                                    ReorderHapticFeedbackType.MOVE
+                                                )
+                                            }
+                                        ),
                                     title = item.space.name.orEmpty(),
                                     icon = item.icon,
                                     previewText = item.previewText,
@@ -803,8 +264,8 @@ fun VaultScreenWithUnreadSection(
                                     onUnmuteSpace = onUnmuteSpace,
                                     onPinSpace = onPinSpace,
                                     onUnpinSpace = onUnpinSpace,
-                                    maxPinnedSpaces = VaultSectionView.MAX_PINNED_SPACES,
                                     onSpaceSettings = onSpaceSettings,
+                                    maxPinnedSpaces = VaultUiState.MAX_PINNED_SPACES,
                                     showPinButton = item.showPinButton
                                 )
                             }
@@ -813,17 +274,31 @@ fun VaultScreenWithUnreadSection(
                                 VaultSpaceCard(
                                     modifier = Modifier
                                         .animateItem()
-                                        .then(
-                                            createCombinedClickableModifier(
-                                                onClick = { onSpaceClicked(item) },
-                                                onLongClick = {
-                                                    expandedSpaceId = item.space.id
-                                                }
-                                            )),
+                                        .graphicsLayer(alpha = alpha.value)
+                                        .combinedClickable(
+                                            onClick = {
+                                                onSpaceClicked(item)
+                                            },
+                                            onLongClick = {
+                                                expandedSpaceId = item.space.id
+                                            }
+                                        )
+                                        .longPressDraggableHandle(
+                                            onDragStarted = {
+                                                hapticFeedback.performHapticFeedback(
+                                                    ReorderHapticFeedbackType.START
+                                                )
+                                            },
+                                            onDragStopped = {
+                                                hapticFeedback.performHapticFeedback(
+                                                    ReorderHapticFeedbackType.MOVE
+                                                )
+                                            }
+                                        ),
                                     title = item.space.name.orEmpty(),
                                     subtitle = item.accessType,
-                                    isPinned = item.isPinned,
                                     icon = item.icon,
+                                    isPinned = item.isPinned,
                                     spaceView = item,
                                     expandedSpaceId = expandedSpaceId,
                                     onDismissMenu = { expandedSpaceId = null },
@@ -831,11 +306,92 @@ fun VaultScreenWithUnreadSection(
                                     onUnmuteSpace = onUnmuteSpace,
                                     onPinSpace = onPinSpace,
                                     onUnpinSpace = onUnpinSpace,
-                                    maxPinnedSpaces = VaultSectionView.MAX_PINNED_SPACES,
+                                    maxPinnedSpaces = VaultUiState.MAX_PINNED_SPACES,
                                     onSpaceSettings = onSpaceSettings,
                                     showPinButton = item.showPinButton
                                 )
                             }
+                        }
+                    }
+                }
+            }
+
+            // Not pinned Spaces Section
+            if (filteredMainSpaces.isNotEmpty()) {
+                itemsIndexed(
+                    items = filteredMainSpaces,
+                    key = { _, item -> item.space.id },
+                    contentType = { _, item ->
+                        when (item) {
+                            is VaultSpaceView.Chat -> TYPE_CHAT
+                            is VaultSpaceView.Space -> TYPE_SPACE
+                        }
+                    }
+                ) { _, item ->
+                    when (item) {
+                        is VaultSpaceView.Chat -> {
+                            VaultChatCard(
+                                modifier = Modifier
+                                    .animateItem()
+                                    .fillMaxWidth()
+                                    .height(80.dp)
+                                    .then(
+                                        createCombinedClickableModifier(
+                                            onClick = { onSpaceClicked(item) },
+                                            onLongClick = { expandedSpaceId = item.space.id }
+                                        ))
+                                    .padding(horizontal = 16.dp),
+                                title = item.space.name.orEmpty(),
+                                icon = item.icon,
+                                previewText = item.previewText,
+                                creatorName = item.creatorName,
+                                messageText = item.messageText,
+                                messageTime = item.messageTime,
+                                chatPreview = item.chatPreview,
+                                unreadMessageCount = item.unreadMessageCount,
+                                unreadMentionCount = item.unreadMentionCount,
+                                attachmentPreviews = item.attachmentPreviews,
+                                isMuted = item.isMuted,
+                                isPinned = item.isPinned,
+                                spaceView = item,
+                                expandedSpaceId = expandedSpaceId,
+                                onDismissMenu = { expandedSpaceId = null },
+                                onMuteSpace = onMuteSpace,
+                                onUnmuteSpace = onUnmuteSpace,
+                                onPinSpace = onPinSpace,
+                                onUnpinSpace = onUnpinSpace,
+                                maxPinnedSpaces = VaultUiState.MAX_PINNED_SPACES,
+                                onSpaceSettings = onSpaceSettings,
+                                showPinButton = item.showPinButton
+                            )
+                        }
+
+                        is VaultSpaceView.Space -> {
+                            VaultSpaceCard(
+                                modifier = Modifier
+                                    .animateItem()
+                                    .then(
+                                        createCombinedClickableModifier(
+                                            onClick = { onSpaceClicked(item) },
+                                            onLongClick = {
+                                                expandedSpaceId = item.space.id
+                                            }
+                                        )),
+                                title = item.space.name.orEmpty(),
+                                subtitle = item.accessType,
+                                isPinned = item.isPinned,
+                                icon = item.icon,
+                                spaceView = item,
+                                expandedSpaceId = expandedSpaceId,
+                                onDismissMenu = { expandedSpaceId = null },
+                                onMuteSpace = onMuteSpace,
+                                onUnmuteSpace = onUnmuteSpace,
+                                onPinSpace = onPinSpace,
+                                onUnpinSpace = onUnpinSpace,
+                                maxPinnedSpaces = VaultUiState.MAX_PINNED_SPACES,
+                                onSpaceSettings = onSpaceSettings,
+                                showPinButton = item.showPinButton
+                            )
                         }
                     }
                 }
