+++ presentation/src/main/java/com/anytypeio/anytype/presentation/vault/Models.kt
@@ -59,15 +59,13 @@ sealed class VaultSpaceView {
     }
 }
 
-/**
- * Data structure for organizing vault spaces into pinned and main sections:
- * @property pinnedSpaces List of pinned spaces (max MAX_PINNED_SPACES)
- * @property mainSpaces List of unpinned spaces
- */
-data class VaultSectionView(
-    val pinnedSpaces: List<VaultSpaceView> = emptyList(),
-    val mainSpaces: List<VaultSpaceView> = emptyList()
-) {
+sealed class VaultUiState {
+    data object Loading : VaultUiState()
+    data class Sections(
+        val pinnedSpaces: List<VaultSpaceView> = emptyList(),
+        val mainSpaces: List<VaultSpaceView> = emptyList()
+    ) : VaultUiState()
+
     companion object {
         const val MAX_PINNED_SPACES = 6
     }
